feat: refactoring the code & add delete api

+++ src/main/java/com/fastcampus/sns/controller/PostController.java
@@ -65,7 +65,7 @@ public class PostController {
 	}
 
 	@GetMapping("/{postId}/likes")
-	public Response<Integer> likeCount(@PathVariable Integer postId) {
+	public Response<Long> likeCount(@PathVariable Integer postId) {
 		return Response.success(postService.likeCount(postId));
 	}
